Fix de la contrasena como null
